% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecsLog.R
\name{logger_api_ecslog_queryBill}
\alias{logger_api_ecslog_queryBill}
\title{写入日志}
\usage{
logger_api_ecslog_queryBill(
  table_name = "t_rapi_ecslog_queryBill",
  FToken = "api_token",
  FBillNo = "abc",
  FCount = 100
)
}
\arguments{
\item{table_name}{表名}

\item{FToken}{api token}

\item{FBillNo}{billno}

\item{FCount}{res count}
}
\value{
return value
}
\description{
写入日志
}
\examples{
logger_api_ecslog_queryBill()
}
