% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecsLog.R
\name{logger_api_ecslog_queryList}
\alias{logger_api_ecslog_queryList}
\title{写入日志}
\usage{
logger_api_ecslog_queryList(
  table_name = "t_rapi_ecslog_queryList",
  FToken = "api_token",
  FStartDate = "2023-04-01",
  FEndDate = "2023-03-31",
  FIsDo = 2,
  FCount = 100
)
}
\arguments{
\item{table_name}{表名}

\item{FToken}{api token}

\item{FStartDate}{start}

\item{FEndDate}{end}

\item{FIsDo}{status}

\item{FCount}{res count}
}
\value{
return value
}
\description{
写入日志
}
\examples{
logger_api_ecslog_queryList()
}
