% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.R
\name{meta_nextId}
\alias{meta_nextId}
\title{查询下一个ID}
\usage{
meta_nextId(table_name = "t_rapi_ecslog_queryList")
}
\arguments{
\item{table_name}{表名}
}
\value{
返回值
}
\description{
查询下一个ID
}
\examples{
meta_maxId()
}
